Create the ssh client interface and implement it
The ssh interface takes as params a command and a stdOut bool. the implementation takes the node index and the ssh port forwarded from dnsmasq. which is the ssh port of the control plane. this is used to later ssh jump from the control plane to the designated target node. it also takes a bool to indicate whether this is a priviliged client or no. the client parses the ssh key and saves it in the struct fields

Signed-off-by: aerosouund <aerosound161@gmail.com>