feat: Create DockerAdapter type
The DockerAdapter type changes the previously used docker.Exec ssh.sh method and conforms it into the SSHClient interface. by taking the command and modifying it to be runnable with ssh.sh based on whether its a script, script with params or direct binary command. the node container field indicates the name of the container that ran the qemu command that corresponds to the target ssh node

Signed-off-by: aerosouund <aerosound161@gmail.com>