refactor: Rewrite run.go to use the SSHClient interface
Use the SSHClientImpl instead of docker exec in run.go, abstract the command to what they actually run instead of the command with ssh.sh or /bin/bash -c. move node running logic to the provisionNode method and the k8s options logic to provisionK8sOptions. leverage the config package to pass configuration to those functions

Signed-off-by: aerosouund <aerosound161@gmail.com>