refactor: Leverage Opts package in run.go and move pure bash flags to gocli flags
Remove the need to run any scripts in the run command and depend on the opts package.
Refactor provisionNode and provisionK8sOptions to use the options pattern instead of passing many args to the constructor.
Add new flags to the gocli which were previously only in the bash code.
Remove prepareDeviceForAssignment and prepareDeviceForAssignment as they both have been moved to opts.

Signed-off-by: aerosouund <aerosound161@gmail.com>