feat/refactor: Introduce the KubevirtProvider type
The KubevirtProvider is a struct representing an arbitrary Kubevirtci running cluster.
It holds all config flags and options that are in the run and provision commands.
A Kubevirt provider can be created in two ways, by creating a cluster using the Start method, or from
an already running cluster.
For this to be possible then json representation of the struct is persisted on the dnsmasq container and later read to parse
the deployed settings
Or through the normal constructor which uses the option pattern to avoid a bloated function signature

The logic that was previously in run.go has been split to several methods to facilitate readability and testing (runNFSGanesha, runRegistry, prepareQemuCmd, prepareDeviceMappings)
and dnsmasq creation logic got moved to its own method instead of existing in its own package

Floating methods such as waitForVMToBeUp, nodeNameFromIndex, nodeContainer.. etc were grouped to be methods of the struct

Signed-off-by: aerosouund <aerosound161@gmail.com>