refactor!: Move provision logic to the KubevirtProvider
Implement provisioning logic on the provider and leverage the linux and k8s provisioning instead of the scripts.
Leverage the same pattern of creating an unpriviliged sshClient to add the ssh key to the root user then use a priviliged client.
Include a map of version major, minor to a version with major, minor and patch to pass it to the provisionk8s opt.

Remove unwanted floating methods such as _cmd, performPhase and implement copyDirectory as a method of the provider type.
Change provision.go command to call the Provision method.
Change type of port flags to uint16 to match the struct type.
Change default value of random-ports to true.

THIS COMMIT BREAKS THE KUBEVIRTCI RELEASE AND VERSIONING PATTERNS

Signed-off-by: aerosouund <aerosound161@gmail.com>